USD Export: shared code to get active uv name.
Added CustomData_get_active_layer_name() function to
BKE_customdata.h to return the name of the active
layer of a given type, as this functionality can
be used in both the USD export and Collada IO
utility functions when retrieving the name of a
mesh's active UV layer.  Refactored the USD Preview
Surface code to use this function.