Compilation of shader_builder.
Shader builder can be used to compile the shader structs at compile
time. This can be activated by enabling the WITH_GPU_SHADER_BUILDER
option.

For compilation we added a lot of stubs as we don't require a full
blender this increases the turn around time when compiling shaders.

Note that we should see if we could split bf_gpu to reduce the amount
of stubs. Or use callback functions reduce the dependencies.