EEVEE: Shadow: Improve the case of directional shadowmaps in ortho view
This is still far from perfect but it is better than not working
correctly.

The view/casters intersection bounds are too big and rough to
compute a decent tilemap level that is near the desire shadow pixels
density.

The algorithm works relatively ok if the sun direction is almost
parallel to the ortho view direction.