Viewport Compositor: Refactor design
This patch refactors the design and architecture of the viewport
compositor.

There is now a clear separation between stages of evaluation of the
compositor setup. Construction, Compilation, Allocation, and Evaluation.
This will allow more granular caching and fit the engine API better.

Instead of dynamic insertion of operations in the operations stream,
operations now have "input processors" that prepares the inputs before
execution, this can be used to do implicit, transform realization, and
any other preprocessing operations that the operation doesn't want to
worry about in its execution method.

Most of the functionalities related to results were moved to that class
and it now does reference counting manually instead of delegating that
to the texture pool.

Other minor refactoring to allow for the above.