Outliner: Display buttons for library overrides
As proposed in T95802, this adds buttons to modify the override in the
library override display mode.
I decided to add the right column so that the left column can be
scrolled independently, like we do it for the restriction toggles in
other display modes.

Contains some cleanups that are to be committed separately.

Differential Revision: https://developer.blender.org/D14268