GPUSource: Add automatic shader nodes GLSL dependency creation
This uses a regex to scan all possible functions definitions and usage at
startup. Replacing the need to maintain a static list of dependency, and
making it compatible with the new shader debugging and include system.

The downside is that it adds quite a bit of startup time (0.5 sec in my tests).
We could mitigate that by either multi-threading some parts, lazily doing
the dependency building, or making it only run at build time.