deps/win: Remove debug symbols from USD
Debug Symbols add a couple of hundred megabytes to the lib
which is already much much larger than I'd like it
to be. None of the other libs ship with symbols, but USD
overrides our C/CXX flags there without an option to bypass.

Updated the patch to forcibly remove this flag.