CMake: allow OpenVDB to find our static TBB libraries
Instead of searching for `${_TBB_LIB_NAME}`, it now also searches for
`${_TBB_LIB_NAME}_static`, which means that OpenVDB and NanoVDB can find
our `libtbb_static.a` files.

Tested on Linux, but since this only adds a file to find and otherwise
keeps the support for the old names as well, shouldn't matter on other
platforms.