GPUMaterial: Rewrite closure eval framework
This remove the split eval function in favor for a more isolated
engine dependent behavior.

Sampled eval engine will be able to select closure at random without
knowing the whole closure weights.

For this we pass the weights alongside the closure properties when it
make sense. When the phenomenon is additive (emission / transparency /
holdout / absorption) the weight is applied to the color directly.