Cycles: Fix missing oneAPI dependencies for kernel compilation
The code was relying on kernel_sources variables set by the CUDA
kernel commands initialization.

Additionally, make sure the output path is the same for both the
custom command itself and for the output file provided to the
add_custom_command. Makes it more clear that no implicit path
deduction is used.