GPUCodegen: Refactor Bump branch derivative support
The previous implementation was duplicating the branch into 3 copies and
created a lot of nodes.

This patch introduce (internal, non user visible) ntree sub functions. This
allows to wrap a portion of the material tree into a function that can be
called multiple times.
This is implemented by `GPU_material_split_sub_function()`.

With the new isolated function we can now define a macro that can compute
the derivatives in an implementation dependent way.
This is implemented by `GPU_differentiate_float_function()`.

The engine then implements `dF_branch(func, result)` the way it wants.
This opens the way to user selectable bump precision per shader.

With this new system a lot of implementation specificity is removed from
`bNode` and the nodetree system.