Clip Editor: Objects popover placement
I have moved the objects popover from the toolbar header to the normal
header. If people would disable the toolbar header they would still need
to be able to see the objects popover.
In Tracking mode the placement is next to the solve error info, which
makes sense, because changing the tracking object also displays a
differen solve error.
In Mask mode it is a little bit weird, because the object popover is now
between the Mask Display popover and the Clip Display popover.