Clip Editor: Clean and Filter Tracks buttons
Add Clean and Filter Tracks Operators to context menu.

Remove Options for the Clean Tracks operator from Cleanup Panel. Filter
Tracks only relies on the Last Operator Menu, so it is a bit weird that
Cleanup has the options visisble and Filter Tracks doesn't.
By removing the options it's consistent for both.