DRW: Simplify shader deferred compilation
Previous method consisted on using another thread and another GL context
to compile in a non blocking way. However this proved to not be a huge
gain and was even making some implementation buggy.

Trying to revert back to a more simplistic approach where we only compile
a handfull of shaders per redraw untill we empty a time budget for shader
compilation.

This approach provides more updates during the compilation process and
avoids all the complexity added by the threading system.
This might be a downside because interleaving many redraw with
compilations could theoreticaly lead to longer wait time until scene wide
compilation is finished. But this still has to be measured.