temp-pbvh-split: Multithread pbvh texture building
I had to use the original threading API for this and
ThreadQueue.  Threads pull nodes to split from a
thread queue and push any new nodes onto the queue
for other threads to further split.

I'm thinking of trying this approach out for PBVH building in
general.  It cut the build time for texture leaves in half.