Viewport Compositor: Fix output storer on Nvidia
The GLSL specification is not clear about passing images to functions and consequently functions
with image parameters are not portable across driver implementations or even non-functioning in
some drivers. See https://github.com/KhronosGroup/GLSL/issues/57.

To work around this, we use macros instead of functions. However, to make those macros usable in
the GPU material library, we also define function counterparts that are guarded with #if 0 such
that they are not used in the shader but are parsed by the GPU shader dependency parser.